fixed delete and create ingredient endpoints, all post man tests pass

+++ server/Services/IngredientsService.cs
@@ -5,14 +5,21 @@ namespace all_spice.Services;
 
 public class IngredientsService
 {
-  public IngredientsService(IngredientsRepository repository)
+  public IngredientsService(IngredientsRepository repository, RecipesService recipesService)
   {
     _repository = repository;
+    _recipesService = recipesService;
   }
   private readonly IngredientsRepository _repository;
+  private readonly RecipesService _recipesService;
 
-  internal Ingredient CreateIngredient(Ingredient ingredientData)
+  internal Ingredient CreateIngredient(Ingredient ingredientData, Account userInfo)
   {
+    Recipe recipe = _recipesService.GetRecipeById(ingredientData.RecipeId);
+    if (recipe.CreatorId != userInfo.Id)
+    {
+      throw new Exception("YOU CANNOT ADD AN INGREDIENT TO SOMEONE ELSES RECIPE");
+    }
     Ingredient ingredient = _repository.CreateIngredient(ingredientData);
     return ingredient;
   }
@@ -35,6 +42,17 @@ public class IngredientsService
 
   internal void DeleteIngredient(int ingredientId, Account userInfo)
   {
+    Ingredient ingredient = GetIngredientById(ingredientId);
+    Recipe recipe = _recipesService.GetRecipeById(ingredient.RecipeId);
+    if (recipe.CreatorId != userInfo.Id)
+    {
+      throw new Exception("YOU CANNOT DELETE ANOTHER PERSONS INGREDIENT");
+    }
+
+    // if (ingredient.Recipe.CreatorId != userInfo.Id)
+    // {
+    //   throw new Exception("YOU CANNOT DELETE ANOTHER PERSONS INGREDIENT");
+    // }
     _repository.DeleteIngredient(ingredientId);
   }
 }
